beamer: move beamer.agent.relayer module to beamer.relayer
